+++ connectors/grok_connect/src/main/java/grok_connect/connectors_info/DbCredentials.java
@@ -19,7 +19,7 @@ public class DbCredentials
     public static final String CONNECTION_STRING = "connString";
     public static final String SSL = "ssl";
     public static final String CACHE_SCHEMA = "cacheSchema";
-    public static final String CACHE_RESULTS = "cache";
+    public static final String CACHE_RESULTS = "cacheResults";
     public static final String CACHE_INVALIDATE_SCHEDULE = "cacheInvalidateSchedule";
     public static final String DB_DESCRIPTION = "Database name";
     public static final String CONNECTION_STRING_DESCRIPTION = "When specified, this connection string overrides " +