+++ js-api/src/entities.ts
@@ -680,6 +680,8 @@ export class FileInfo extends Entity {
     super(dart);
   }
 
+  get connection(): DataConnection { return api.grok_FileInfo_Get_Connection(toJs(this.dart)); }
+
   /** Returns path, i.e. `geo/dmv_offices.csv` */
   get path(): string { return api.grok_FileInfo_Get_Path(this.dart); }
 
@@ -1199,6 +1201,9 @@ export interface PropertyOptions {
   fieldName?: string;
 
   tags?: any;
+
+  /** Filter for columns, can be numerical, categorical or directly a column type (string, int...) */
+  columnTypeFilter?: ColumnType | 'numerical' | 'categorical';
 }
 
 
@@ -1356,10 +1361,7 @@ export class Property {
    * It is editable via the context panel, and gets saved into the view layout as well.
    * Property getter/setter typically uses Widget's "temp" property for storing the value. */
   static registerAttachedProperty(typeName: string, property: Property) {
-    throw 'Not implemented';
-    // Andrew: looks like my commit got lost somewhere :(
-    // Will need to bring it back, it was a nice feature
-    // api.grok_Property_RegisterAttachedProperty(typeName, property.dart);
+    api.grok_Property_RegisterAttachedProperty(typeName, property.dart);
   }
 }