+++ packages/ApiTests/queries/dummy-query.sql
@@ -1,5 +1,5 @@
 --name: dummyPackageQuery
 --connection: System:Datagrok
---input: int x
+--input: double x
 SELECT @x as res;
 --end