+++ packages/EDA/wasm/regression.h
@@ -1,9 +1,6 @@
 // regression.h
-
-/* Linear regression coefficients computation.
-   
-   Implemented 2 approaches: with & without data normalizing.
-   Data normalizing provides better results in te case of big data. */
+// 
+//Implementation linear regression coefficients computation
 
 #include "../../../../Eigen/Eigen/Dense"
 using namespace Eigen;