+++ packages/HitTriage/src/app/hit-design-app.ts
@@ -148,6 +148,33 @@ export class HitDesignApp extends HitAppBase<HitDesignTemplate> {
 
   set campaign(campaign: HitDesignCampaign | undefined) {this._campaign = campaign;}
 
+  private _duplicateVidCache?: {
+    colVersion: number,
+    valueCounts: Int32Array | Uint32Array,
+    indexes: Int32Array | Uint32Array,
+  };
+
+  private cacheDuplicateVIDs() {
+    const col = this.dataFrame?.col(ViDColName);
+    if (!col) {
+      this._duplicateVidCache = undefined;
+      return;
+    }
+    const colVersion = col.version;
+    const valueCounts = new Uint32Array(col.categories.length).fill(0);
+    const indexes = col.getRawData() as Uint32Array;
+    indexes.forEach((v) => valueCounts[v]++);
+    this._duplicateVidCache = {colVersion, valueCounts, indexes};
+  }
+
+  get duplicateVidCache() {
+    const col = this.dataFrame?.col(ViDColName);
+    if (!col)
+      return undefined;
+    if (!this._duplicateVidCache || col.version !== this._duplicateVidCache.colVersion)
+      this.cacheDuplicateVIDs();
+    return this._duplicateVidCache;
+  }
   private getDesignView(): DG.TableView {
     const subs: Subscription[] = [];
     const isNew = this.dataFrame!.col(this.molColName)?.toList().every((m) => !m && m === '');
@@ -155,6 +182,29 @@ export class HitDesignApp extends HitAppBase<HitDesignTemplate> {
     this._designViewName = this.campaign?.name ?? this._designViewName;
     view.name = this._designViewName;
     this.processedValues = this.dataFrame!.getCol(this.molColName).toList();
+
+    const performSingleCellCalculations = async (newValueIdx: number, newValue?: string) => {
+      const computeObj = this.template!.compute;
+      if (!newValue || newValue === '')
+        return;
+
+      const calcDf =
+              await calculateSingleCellValues(
+                newValue, computeObj.descriptors.args, computeObj.functions, computeObj.scripts, computeObj.queries);
+
+      for (const col of calcDf.columns.toList()) {
+        if (col.name === HitDesignMolColName) continue;
+        if (!this.dataFrame!.columns.contains(col.name)) {
+          const newCol = this.dataFrame!.columns.addNew(col.name, col.type);
+          newCol.semType = col.semType;
+        }
+            this.dataFrame!.col(col.name)!.set(newValueIdx, col.get(0), false);
+      }
+          this.dataFrame!.fireValuesChanged();
+          this.saveCampaign(undefined, false);
+    };
+
+
     setTimeout(async () => {
       view._onAdded();
       await new Promise((r) => setTimeout(r, 1000)); // needed for substruct filter
@@ -222,7 +272,7 @@ export class HitDesignApp extends HitAppBase<HitDesignTemplate> {
             return;
           if (!viewer.tableView || viewer.tableView.id !== view.id)
             return;
-          if (args?.args?.item?.tableColumn?.name !== this.molColName)
+          if (args?.args?.item?.tableColumn?.name !== this.molColName || !args?.args?.item?.isTableCell)
             return;
           const menu: DG.Menu = args?.args?.menu;
           if (!menu)
@@ -248,6 +298,18 @@ export class HitDesignApp extends HitAppBase<HitDesignTemplate> {
               console.error(e);
             }
           });
+
+          const cellIndex = args?.args?.item?.tableRowIndex;
+          const cellValue = args?.args?.item?.cell?.value;
+          if (cellValue && (cellIndex ?? -1) > -1) {
+            menu.item('Re-Run Calculations', async () => {
+              try {
+                await performSingleCellCalculations(cellIndex, cellValue);
+              } catch (e) {
+                console.error(e);
+              }
+            });
+          }
         } catch (e: any) {
           grok.log.error(e);
         }
@@ -262,34 +324,68 @@ export class HitDesignApp extends HitAppBase<HitDesignTemplate> {
             return;
           const newValue = gc.cell.value;
           const newValueIdx = gc.tableRowIndex!;
-
-          if (!this.dataFrame!.col(ViDColName)?.get(newValueIdx) ||
-            this.dataFrame!.col(ViDColName)?.get(newValueIdx) === '')
-              this.dataFrame!.col(ViDColName)!.set(newValueIdx, getNewVid(this.dataFrame!.col(ViDColName)!), false);
-
-
-          const computeObj = this.template!.compute;
-          if (!newValue || newValue === '')
-            return;
-
-          const calcDf =
-              await calculateSingleCellValues(
-                newValue, computeObj.descriptors.args, computeObj.functions, computeObj.scripts, computeObj.queries);
-
-          for (const col of calcDf.columns.toList()) {
-            if (col.name === HitDesignMolColName) continue;
-            if (!this.dataFrame!.columns.contains(col.name)) {
-              const newCol = this.dataFrame!.columns.addNew(col.name, col.type);
-              newCol.semType = col.semType;
+          let newVid = this.dataFrame!.col(ViDColName)?.get(newValueIdx);
+          let foundMatch = false;
+          // try to find existing molecule
+          if (newValue) {
+            try {
+              const canonicals = gc.tableColumn.toList().map((cv) => {
+                try {
+                  return grok.chem.convert(cv, grok.chem.Notation.Unknown, grok.chem.Notation.Smiles);
+                } catch (e) {
+                  return '';
+                }
+              },
+              );
+              const canonicalNewValue =
+                grok.chem.convert(newValue, grok.chem.Notation.Unknown, grok.chem.Notation.Smiles);
+              if (canonicals?.length === this.dataFrame!.rowCount) {
+                for (let i = 0; i < canonicals.length; i++) {
+                  if (canonicals[i] === canonicalNewValue &&
+                      i !== newValueIdx && this.dataFrame!.col(ViDColName)?.get(i)) {
+                    newVid = this.dataFrame!.col(ViDColName)?.get(i);
+                    foundMatch = true;
+                    break;
+                  }
+                }
+              }
+            } catch (e) {
+              console.error(e);
             }
-            this.dataFrame!.col(col.name)!.set(newValueIdx, col.get(0), false);
           }
-          this.dataFrame!.fireValuesChanged();
-          this.saveCampaign(undefined, false);
+          // if the vid was duplicated, generate a new one
+          if (this.duplicateVidCache && !foundMatch &&
+            this.duplicateVidCache.valueCounts[this.duplicateVidCache.indexes[newValueIdx]] > 1)
+            newVid = null;
+
+          if (!newVid || newVid === '')
+            newVid = getNewVid(this.dataFrame!.col(ViDColName)!);
+
+          this.dataFrame!.col(ViDColName)!.set(newValueIdx, newVid, false);
+
+          performSingleCellCalculations(newValueIdx, newValue);
         } catch (e) {
           console.error(e);
         }
       }));
+
+      subs.push(view.grid.onCellRender.subscribe((args) => {
+        try {
+          // color duplicate vid values
+          const cell = args.cell;
+          if (!cell || !cell.isTableCell || !cell.tableColumn || !this.duplicateVidCache ||
+            cell.tableColumn.name !== ViDColName || (cell.tableRowIndex ?? -1) < 0)
+            return;
+
+          if (this.duplicateVidCache.valueCounts[this.duplicateVidCache.indexes[cell.tableRowIndex!]] > 1) {
+            args.cell.style.backColor =
+              DG.Color.setAlpha(DG.Color.getCategoricalColor(this.duplicateVidCache.indexes[cell.tableRowIndex!])
+                , 150);
+          }
+        } catch (e) {}
+      }));
+
+
       const onRemoveSub = grok.events.onViewRemoved.subscribe((v) => {
         if (v.id === view.id) {
           subs.forEach((s) => s.unsubscribe());