+++ packages/Peptides/src/package.ts
@@ -14,6 +14,7 @@ import {PeptidesModel} from './model';
 import {macromoleculeSarFastaDemoUI} from './demo/fasta';
 import {u2} from '@datagrok-libraries/utils/src/u2';
 import {ClusterMaxActivityViewer} from './viewers/cluster-max-activity-viewer';
+import {LSTPieChartRenderer} from './utils/cell-renderer';
 
 let monomerWorks: MonomerWorks | null = null;
 let treeHelper: ITreeHelper;
@@ -50,7 +51,6 @@ async function openDemoData(chosenFile: string): Promise<void> {
 }
 
 //name: Peptides
-//tags: app
 //output: view v
 export function Peptides(): DG.View {
   const appHeader = u2.appHeader({
@@ -168,4 +168,13 @@ export function manualAlignment(_monomer: string): DG.Widget {
 //meta.isDemoScript: False
 export async function macromoleculeSarFastaDemo(): Promise<void> {
   return macromoleculeSarFastaDemoUI();
+}
+
+//name: LST Pie Chart
+//tags: cellRenderer
+//meta.cellType: lst-pie-chart
+//meta.gridChart: true
+//output: grid_cell_renderer result
+export function lstPiechartCellRenderer(): LSTPieChartRenderer {
+  return new LSTPieChartRenderer();
 }