+++ libraries/utils/CHANGELOG.md
@@ -1,5 +1,21 @@
 # utils changelog
 
+## 4.2.15 (2024-07-30)
+
+Removed Unhadled exception test for benchmarks 
+
+## 4.2.14 (2024-07-29)
+
+Fixed setting lines width, color and opacity in ScatterPlotLinesRenderer
+
+## 4.2.13 (2024-07-15)
+
+Added benchmark flag to category options
+
+## 4.2.12 (2024-07-15)
+
+Added benchmark flag to test options
+
 ## 4.2.11 (2023-06-20)
 
 New tests reporting method.